requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
dnspython==2.6.1
fastapi==0.115.0
fastapi-csrf-protect==0.3.6
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
motor==3.6.0
packaging==24.1
passlib==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.9.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
